+++ sdk/core/azure-core/inc/azure/core/exception.hpp
@@ -8,7 +8,12 @@
 
 #pragma once
 
+#include "azure/core/http/http_status_code.hpp"
+#include "azure/core/http/raw_response.hpp"
+
+#include <memory>
 #include <stdexcept>
+#include <string>
 
 namespace Azure { namespace Core {
   /**
@@ -18,10 +23,55 @@ namespace Azure { namespace Core {
   class RequestFailedException : public std::runtime_error {
   public:
     /**
+     * @brief The Http response code.
+     *
+     */
+    Azure::Core::Http::HttpStatusCode StatusCode = Azure::Core::Http::HttpStatusCode::None;
+
+    /**
+     * @brief The Http reason phrase from the response.
+     *
+     */
+    std::string ReasonPhrase;
+
+    /**
+     * @brief The client request header from the Http response.
+     *
+     */
+    std::string ClientRequestId;
+
+    /**
+     * @brief The request id header from the Http response.
+     *
+     */
+    std::string RequestId;
+
+    /**
+     * @brief The error code from service returned in the Http response.
+     *
+     */
+    std::string ErrorCode;
+
+    /**
+     * @brief The error message from the service returned in the Http response.
+     *
+     */
+    std::string Message;
+
+    /**
+     * @brief The entire Http raw response.
+     *
+     */
+    std::unique_ptr<Azure::Core::Http::RawResponse> RawResponse;
+
+    /**
      * @brief Construct a new Request Failed Exception object.
      *
      * @param message The error description.
      */
-    explicit RequestFailedException(std::string const& message) : std::runtime_error(message) {}
+    explicit RequestFailedException(std::string const& message)
+        : std::runtime_error(message), Message(message)
+    {
+    }
   };
 }} // namespace Azure::Core
