+++ sdk/core/azure-core/src/http/raw_response.cpp
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
+#include "azure/core/http/raw_response.hpp"
 #include "azure/core/http/http.hpp"
 #include "azure/core/internal/strings.hpp"
